+++ src/data/sort/primitives.rs
@@ -93,7 +93,7 @@ macro_rules! impl_sort {
                 let mut arr = self.0;
                 cfor![i in 0..N => {
                     cfor![j in 0..N-i-1 => {
-                        iif![arr[j] > arr[j+1]; cswap!(xor arr[j], arr[j+1])];
+                        iif![arr[j] > arr[j+1]; cswap!(xor: arr[j], arr[j+1])];
                     }];
                 }];
                 arr
@@ -106,7 +106,7 @@ macro_rules! impl_sort {
                 cfor![i in 1..N => {
                     let mut j = i;
                     while j > 0 && arr[j-1] > arr[j] {
-                        cswap!(xor arr[j], arr[j-1]);
+                        cswap!(xor: arr[j], arr[j-1]);
                         j -= 1;
                     }
                 }];
@@ -122,7 +122,7 @@ macro_rules! impl_sort {
                     cfor![j in (i+1)..N => {
                         iif![arr[j] < arr[min_index]; min_index = j];
                     }];
-                    cswap!(xor arr[min_index], arr[i]);
+                    cswap!(xor: arr[min_index], arr[i]);
                 }];
                 arr
             }
@@ -139,7 +139,7 @@ macro_rules! impl_sort {
                 let mut arr = self.0;
                 cfor![i in 0..N => {
                     cfor![j in 0..N-i-1 => {
-                        iif![Compare(arr[j]).gt(arr[j+1]); cswap!(arr[j], arr[j+1])];
+                        iif![Compare(arr[j]).gt(arr[j+1]); cswap!(tmp: arr[j], arr[j+1])];
                     }];
                 }];
                 arr
@@ -152,7 +152,7 @@ macro_rules! impl_sort {
                 cfor![i in 1..N => {
                     let mut j = i;
                     while j > 0 && Compare(arr[j-1]).gt(arr[j]) {
-                        cswap!(arr[j], arr[j-1]);
+                        cswap!(tmp: arr[j], arr[j-1]);
                         j -= 1;
                     }
                 }];
@@ -168,7 +168,7 @@ macro_rules! impl_sort {
                     cfor![j in (i+1)..N => {
                         iif![Compare(arr[j]).lt(arr[min_index]); min_index = j];
                     }];
-                    cswap!(arr[min_index], arr[i]);
+                    cswap!(tmp: arr[min_index], arr[i]);
                 }];
                 arr
             }